Refactored into smaller components & styles
