+++ src/components/details/EventTile.scss
@@ -16,7 +16,7 @@
     justify-content: space-between;
 
     font-weight: 500;
-    font-size: 3em;
+    font-size: 2em;
     padding-bottom: 0.5em;
   }
 
@@ -25,7 +25,7 @@
     flex-direction: row;
     justify-content: space-between;
 
-    font-size: 3em;
+    font-size: 2em;
     font-weight: 400;
 
     padding-bottom: 0.5em;
@@ -33,7 +33,7 @@
 
   &-description {
     font-weight: 200;
-    font-size: 1.8em;
+    font-size: 1em;
   }
 
   &-favorites-svg {
