+++ src/styles.scss
@@ -1,12 +1,13 @@
 @import "./src/components/list/EventList.scss";
-@import "./src/components/details/EventDetails.scss";
+@import "./src/components/details/EventTile.scss";
 @import "/src/components/menu/Menu.scss";
 
 * {
   font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif;
 }
 
-html, body {
+html,
+body {
   background-color: #262626;
   width: 100vw;
   height: 100vh;
